moved pages into folders for CSS
