+++ src/pages/ProfileDisplay/Profile.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import { Link } from "react-router-dom";
-import { getLocations, addLocation } from "../services/locations";
+import { getLocations, addLocation } from "../../services/locations";
 
 const Profile = () => {
   const [locations, setLocations] = useState([]);
